+++ src/pages/Login/index.jsx
@@ -1,48 +1,8 @@
 import React from "react";
-import { GoogleAuthProvider, getAuth, signInWithPopup } from "firebase/auth";
-import { useNavigate } from "react-router-dom";
 import NavBar from "../../components/NavBar";
+import Button from "../../components/Button";
 
 const Login = () => {
-  const navigate = useNavigate();
-  const provider = new GoogleAuthProvider();
-  const auth = getAuth();
-  const signInwithGoogle = async () => {
-    try {
-      const result = await signInWithPopup(auth, provider);
-      const user = result.user;
-
-      if (user) {
-        const tkn = await user.getIdToken();
-
-        sessionStorage.setItem("accessToken", tkn);
-
-        const fetchData = async (tkn) => {
-          try {
-            const response = await fetch("http://localhost:3000/login", {
-              method: "POST",
-              headers: {
-                Authorization: tkn,
-              },
-            });
-
-            if (response.ok) {
-              navigate("/");
-            } else {
-              throw new Error("요청이 실패했습니다");
-            }
-          } catch (error) {
-            console.error(error);
-          }
-        };
-
-        await fetchData(tkn);
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
   return (
     <>
       <div className="flex flex-col space-y-40">
@@ -57,16 +17,9 @@ const Login = () => {
                 Welcome Back
               </h2>
             </div>
-
             <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-sm">
               <div>
-                <button
-                  type="submit"
-                  className="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-                  onClick={signInwithGoogle}
-                >
-                  sign in with Google
-                </button>
+                <Button />
               </div>
             </div>
           </div>
